requirements.txt
alembic==1.14.1
babel==2.17.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
